added scraper class
